Refactoring after rebasing (separated ExecutionContexts per type introduced)
