--- src/main/java/org/openhab/automation/jrule/internal/engine/JRuleValueComparators.java
@@ -1,16 +1,0 @@
-package org.openhab.automation.jrule.internal.engine;
-
-public interface JRuleValueComparators {
-
-    public Double getGt();
-
-    public Double getGte();
-
-    public Double getLt();
-
-    public Double getLte();
-
-    public String getEq();
-
-    public String getNeq();
-}
